Decompose the 268-LOC build_worker_cos_statuses_from_maps body (formerly at worker/cos.rs:535-832) into three sibling submodules under the new worker/cos/ directory module:
  - cos/interface_row.rs::accumulate_interface_root  (interface-level
    fields: ifindex, name resolution, shaping/burst MAX, worker_instances,
    timer-wheel sleeper sums)
  - cos/queue_row.rs::accumulate_queue_row           (queue-scoped fields:
    priority gate, exact/guarantee, transmit rate, queued bytes,
    runnable/parked, surplus/deficit, #784 active_flow_buckets_peak MAX,
    flow_fair, #710/#718 drop counters, #751 per-queue drain telemetry,
    #760 overshoot-hunt counters)
  - cos/status.rs::build_worker_cos_statuses_from_maps  (orchestrator
    over (cos_map, binding_live) pairs; #709/#748/#751 binding-scoped
    merge stays inline as a 3-line gated if per Gemini round-1 KILL
    counter-example) plus finalize_interface_vec (Vec flatten +
    nonempty/runnable counts + sort by (interface_name, ifindex))

Pure code motion of the existing block. Every saturating_add / MAX,
every Relaxed atomic load, the priority gate (status.worker_instances
== 0 read BEFORE increment), the drain_latency_hist.resize gate
(queue_invocations > 0), and the unique-owner-local-exact-queue
binding-scoped attribution rule are preserved verbatim.

The pure git mv from cos.rs/cos_tests.rs to cos/mod.rs/cos/tests.rs
landed in the prior commit (1fd1c62a -> ...). This commit:

  - moves the 268-LOC fn body out of cos/mod.rs into the new
    cos/{interface_row,queue_row,status}.rs files,
  - drops the standalone build_worker_cos_statuses wrapper from
    cos/mod.rs into cos/status.rs,
  - re-exports the orchestrator entry at cos/mod.rs scope so worker/
    mod.rs's existing 'use cos::build_worker_cos_statuses' line keeps
    resolving without churn,
  - adds the #784 MAX-not-sum regression test
    (active_flow_buckets_peak_is_max_not_sum_across_workers) per Codex
    round-1 PLAN-NEEDS-MAJOR finding 3.

Plan + adversarial review under docs/pr/1349-worker-cos-status-split/:
  - AGY plan-review:    PLAN-READY (no findings)
  - Claude SMR review:  PLAN-READY
  - Codex round-3:      PLAN-READY (after r1 MAJOR + r2 MAJOR + r3 MINOR
                        addressed across plan v1 -> v3)
  - Gemini round-2:     PLAN-KILL (style/linear-mapper objection;
                        not blocking per feedback_gemini_low_signal_on_refactor)

Validation:
  - cargo build --release: clean
  - cargo test --release --bin xpf-userspace-dp: 1434/1434 pass
  - cargo test afxdp::worker::cos: 24/24 pass (includes new #784 test)
  - 5x flake on build_worker_cos_statuses_owner_profile_only_surfaces_
    on_unambiguous_owner_local_exact_queue: 5/5 pass

Closes #1349
